+++ src/frequenz/channels/experimental/__init__.py
@@ -10,11 +10,11 @@ Warning:
 """
 
 from ._pipe import Pipe
-from ._predicates import OnlyIfPrevious
+from ._predicates import WithPrevious
 from ._relay_sender import RelaySender
 
 __all__ = [
-    "OnlyIfPrevious",
+    "WithPrevious",
     "Pipe",
     "RelaySender",
 ]